Refactor init functions to impl level
