+++ import-export-cli/cmd/init.go
@@ -19,25 +19,11 @@
 package cmd
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
-	"unicode"
 
-	"github.com/Jeffail/gabs"
-
-	jsoniter "github.com/json-iterator/go"
-	"github.com/wso2/product-apim-tooling/import-export-cli/box"
-
-	yaml2 "gopkg.in/yaml.v2"
-
-	"github.com/go-openapi/loads"
-	v2 "github.com/wso2/product-apim-tooling/import-export-cli/specs/v2"
-
-	"github.com/ghodss/yaml"
+	"github.com/wso2/product-apim-tooling/import-export-cli/impl"
 
 	"github.com/spf13/cobra"
 
@@ -56,252 +42,6 @@ const initCmdExample = `apictl init myapi --oas petstore.yaml
 apictl init Petstore --oas https://petstore.swagger.io/v2/swagger.json
 apictl init Petstore --oas https://petstore.swagger.io/v2/swagger.json --initial-state=PUBLISHED
 apictl init MyAwesomeAPI --oas ./swagger.yaml -d definition.yaml`
-
-// directories to be created
-var dirs = []string{
-	utils.InitProjectDefinitions,
-	utils.InitProjectImage,
-	utils.InitProjectDocs,
-	utils.InitProjectSequences,
-	utils.InitProjectSequencesFault,
-	utils.InitProjectSequencesIn,
-	utils.InitProjectSequencesOut,
-	utils.InitProjectClientCertificates,
-	utils.InitProjectClientCertificates,
-	utils.InitProjectInterceptors,
-	utils.InitProjectLibs,
-}
-
-// createDirectories will create dirs in current working directory
-func createDirectories(name string) error {
-	for _, dir := range dirs {
-		dirPath := filepath.Join(name, filepath.FromSlash(dir))
-		utils.Logln(utils.LogPrefixInfo + "Creating directory " + dirPath)
-		err := os.MkdirAll(dirPath, os.ModePerm)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// loadDefaultSpecFromDisk loads api definition stored in HOME/.wso2apictl/default_api.yaml
-func loadDefaultSpecFromDisk() (*v2.APIDefinitionFile, error) {
-	defaultData, err := ioutil.ReadFile(utils.DefaultAPISpecFilePath)
-	if err != nil {
-		return nil, err
-	}
-	def := &v2.APIDefinitionFile{}
-	err = yaml.Unmarshal(defaultData, &def)
-	if err != nil {
-		return nil, err
-	}
-	return def, nil
-}
-
-// loads swagger from swaggerDoc
-// swagger2.0/OpenAPI3.0 specs are supported
-func loadSwagger(swaggerDoc string) (*loads.Document, error) {
-	utils.Logln(utils.LogPrefixInfo + "Loading swagger from " + swaggerDoc)
-	return loads.Spec(swaggerDoc)
-}
-
-// hasJSONPrefix returns true if the provided buffer appears to start with
-// a JSON open brace.
-func hasJSONPrefix(buf []byte) bool {
-	return hasPrefix(buf, []byte("{"))
-}
-
-// Return true if the first non-whitespace bytes in buf is prefix.
-func hasPrefix(buf []byte, prefix []byte) bool {
-	trim := bytes.TrimLeftFunc(buf, unicode.IsSpace)
-	return bytes.HasPrefix(trim, prefix)
-}
-
-// executeInitCmd will run init command
-func executeInitCmd() error {
-	var dir string
-	swaggerSavePath := filepath.Join(initCmdOutputDir, filepath.FromSlash("Definitions/swagger.yaml"))
-
-	if initCmdOutputDir != "" {
-		err := os.MkdirAll(initCmdOutputDir, os.ModePerm)
-		if err != nil {
-			return err
-		}
-		p, err := filepath.Abs(initCmdOutputDir)
-		if err != nil {
-			return err
-		}
-		dir = p
-	} else {
-		pwd, err := os.Getwd()
-		if err != nil {
-			return err
-		}
-		dir = pwd
-	}
-	fmt.Println("Initializing a new WSO2 API Manager project in", dir)
-
-	definitionFile, err := loadDefaultSpecFromDisk()
-
-	// Get the API DTO specific details to process
-	def := &definitionFile.Data
-	if err != nil {
-		return err
-	}
-
-	// initCmdInitialState has already validated before creating the 'dir'
-	if initCmdInitialState != "" {
-		def.LifeCycleStatus = initCmdInitialState
-	}
-
-	err = createDirectories(initCmdOutputDir)
-	if err != nil {
-		return err
-	}
-
-	// use swagger to auto generate
-	if initCmdSwaggerPath != "" {
-		// load swagger from path
-		doc, err := loadSwagger(initCmdSwaggerPath)
-		if err != nil {
-			return err
-		}
-		// We use swagger2 loader. It works fine for now
-		// Since we don't use 3.0 specific details its ok
-		// otherwise please use v2.openAPI3 loaders
-		err = v2.Swagger2Populate(def, doc)
-		if err != nil {
-			return err
-		}
-
-		// convert and save swagger as yaml
-		yamlSwagger, err := utils.JsonToYaml(doc.Raw())
-		if err != nil {
-			return err
-		}
-
-		// write to file
-		err = ioutil.WriteFile(swaggerSavePath, yamlSwagger, os.ModePerm)
-		if err != nil {
-			return err
-		}
-	} else {
-		// create an empty swagger
-		utils.Logln(utils.LogPrefixInfo + "Writing " + swaggerSavePath)
-		swaggerDoc, _ := box.Get("/init/swagger-default.yaml")
-		err = ioutil.WriteFile(swaggerSavePath, swaggerDoc, os.ModePerm)
-		if err != nil {
-			return err
-		}
-	}
-
-	// use api definition if given
-	if initCmdApiDefinitionPath != "" {
-		// read definition file
-		utils.Logln(utils.LogPrefixInfo + "Reading API Definition from " + initCmdApiDefinitionPath)
-		content, err := ioutil.ReadFile(initCmdApiDefinitionPath)
-		if err != nil {
-			return err
-		}
-
-		apiDef := &v2.APIDefinitionFile{}
-
-		// substitute env variables
-		utils.Logln(utils.LogPrefixInfo + "Substituting environment variables")
-		data, err := utils.EnvSubstitute(string(content))
-		if err != nil {
-			return err
-		}
-		content = []byte(data)
-
-		// read from yaml definition
-		err = yaml2.Unmarshal(content, &apiDef)
-		if err != nil {
-			return err
-		}
-
-		// marshal original def
-		originalDefBytes, err := jsoniter.Marshal(definitionFile)
-		if err != nil {
-			return err
-		}
-		// marshal new def
-		newDefBytes, err := jsoniter.Marshal(apiDef)
-		if err != nil {
-			return err
-		}
-
-		// merge two definitions
-		finalDefBytes, err := utils.MergeJSON(originalDefBytes, newDefBytes)
-		if err != nil {
-			return err
-		}
-		tmpDef := &v2.APIDefinitionFile{}
-		err = json.Unmarshal(finalDefBytes, &tmpDef)
-		if err != nil {
-			return err
-		}
-		definitionFile.Data = tmpDef.Data
-	}
-
-	apiData, err := yaml2.Marshal(definitionFile)
-	if err != nil {
-		return err
-	}
-
-	// write to the disk
-	apiJSONPath := filepath.Join(initCmdOutputDir, filepath.FromSlash("api.yaml"))
-	utils.Logln(utils.LogPrefixInfo + "Writing " + apiJSONPath)
-	err = ioutil.WriteFile(apiJSONPath, apiData, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	apimProjDeploymentEnvironmentsFilePath := filepath.Join(initCmdOutputDir, "deployment_environments.yaml")
-	utils.Logln(utils.LogPrefixInfo + "Writing " + apimProjDeploymentEnvironmentsFilePath)
-	deploymentEnvironments, _ := box.Get("/init/default_deployment_environments.yaml")
-	err = ioutil.WriteFile(apimProjDeploymentEnvironmentsFilePath, deploymentEnvironments, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	apimProjReadmeFilePath := filepath.Join(initCmdOutputDir, "README.md")
-	utils.Logln(utils.LogPrefixInfo + "Writing " + apimProjReadmeFilePath)
-	readme, _ := box.Get("/init/README.md")
-	err = ioutil.WriteFile(apimProjReadmeFilePath, readme, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	// Create metaData struct using details from definition
-	metaData := utils.MetaData{
-		Name:    definitionFile.Data.Name,
-		Version: definitionFile.Data.Version,
-	}
-	marshaledData, err := jsoniter.Marshal(metaData)
-	if err != nil {
-		return err
-	}
-
-	jsonMetaData, err := gabs.ParseJSON(marshaledData)
-	metaDataContent, err := utils.JsonToYaml(jsonMetaData.Bytes())
-	if err != nil {
-		return err
-	}
-
-	// write api_meta.yaml file to the project directory
-	apiMetaDataPath := filepath.Join(initCmdOutputDir, filepath.FromSlash(utils.MetaFileAPI))
-	utils.Logln(utils.LogPrefixInfo + "Writing " + apiMetaDataPath)
-	err = ioutil.WriteFile(apiMetaDataPath, metaDataContent, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	fmt.Println("Project initialized")
-	fmt.Println("Open README file to learn more")
-	return nil
-}
 
 var InitCommand = &cobra.Command{
 	Use:     "init [project path]",
@@ -344,7 +84,7 @@ var InitCommand = &cobra.Command{
 			}
 		}
 
-		err := executeInitCmd()
+		err := impl.InitAPIProject(initCmdOutputDir, initCmdInitialState, initCmdSwaggerPath, initCmdApiDefinitionPath, false)
 		if err != nil {
 			utils.HandleErrorAndContinue("Error initializing project", err)
 			// Remove the already created project with its content since it is partially created and wrong
